Force 10-digit phone numbers (do not allow 7-digit local calls).
Reformat the phone number when it is valid to match the nnn-nnn-nnnn
format.

If the number is valid and then is made invalid, Angular is still
calling the ng-change handler. Annoying. We really want to pass a
reference to the model we want to examine so our "onBlur" handler
can reformat and set the value only when the input/model is valid.

This likely needs a custom directive.

Ignoring the error, this does what we want it to do...