Fixes https://trello.com/c/rmCHgWj8
EPA and State waste codes and report management method codes are sent to the
server and stored as arrays of values.

The display is not the best on the manifest detail page; angular is
automagically joining the values together with commas and no spaces and
constructs lengthy non-breaking strings. Presentation issues will be dealt with
later.

Fix some typos along the way.