Sign a manifest using the /api/manifest/sign API.
We appear to be able to sign a manifest round-trip from UI to API to CDX.

Start the signing process by going to /web/sign.html#!/?id={emanifest id}
and enter your CROMERR username/password. Then answer the security question.
Upon submission, the manifest's JSON representation will be transmitted
as a document in a CROMERR signing service.

Time to do some cleanup and wire the page flow together. There's no way
for the user to enter the signing process nor is there any indication of
success or failure.