Use the template engine for all pages. Some were not using templates.
Fix a small bug found where things would fail in search if the data in
the manifests was not valid. Since we have no validation enforcing
required content in the manifest submission, we might as well be
consistent with that expectation in other places.

Fixes https://trello.com/c/gjgDGz6Z