Add the discovery header to the rest of the site (previously it was on the demo web interface for manifest submission and retrieval).
It's a bit of a hybrid solution-- we bring in the web standards css
(from https://github.com/18F/web-design-standards) and override
some of the styles that changed the existing site too much. The
discovery header really shouldn't be where it wound up, but it
would have made more changes to the DOM or style overriding if
we put it in the right place. The clean solution is to recast the
API documentation and demonstration and home parts of the site
with the web standards in mind. That's a bigger lift.

Fixes #7. Fixes https://trello.com/c/0JnAifym