Modify the US Web Design Standards implementation to match the comp for the industry manifest entry form as closely as possible.
We do not have the glyphs for add/remove or style the add/remove text
as requested.

We do not have the smaller inputs for container type and unit so
they're not side by side with their related inputs. We can't easily
style the select option to display a particular value when selected
and a different value when the select is open /and/ make it readily
accessible.