Fix deployments to cloud.gov.
Apparently one of the build packs we use originally moved the database.yml
file out of the way in lieu of connecting to the database as

`Binding service e-manifest-dev-db to app e-manifest in org epa-emanifest / space prod`

Now it doesn't muck with the database.yml and the database.yml gets in
the way. We now ignore the database.yml file in our .cfignore to prevent
the file from being pushed to cloud.gov.

Huge thanks to @dlapiduz for figuring it all out.