Wrap views in usa grid classes
To prevent a small gap of blue between the main content and the nav
content we need to wrap the views in the `usa-width-*` classes. We
can’t do this at the `application.html.erb` level because we need the
views to be able to configure themselves as two-thirds/one-third or
full width.