Update _manifest.json
Since this is an example, I moved the one mandatory field (manifest_tracking_number) to the top so that people testing the app can easily find it and edit it.  I thought about changing the value in the example for "manifest_tracking_number" to a note that would not validate, but decided not to. here is an example of that thought:

      "manifest_tracking_number" : "ENTER A STRING OF 9 Digits and 3 Character, for example 123456789ABC"

Would this be useful or should we leave it as an example.